Move resource creation, readbacks, and InvocationState from DX backend to shared file.
