+++ lib/Support/CMakeLists.txt
@@ -1,1 +1,1 @@
-add_offloadtest_library(Support Pipeline.cpp Check.cpp)
+add_offloadtest_library(Support Pipeline.cpp Check.cpp OffloadMigration.cpp)
